+++ src/pytorch_lightning/CHANGELOG.md
@@ -224,7 +224,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Removed the deprecated `SimpleProfiler.profile_iterable` and `AdvancedProfiler.profile_iterable` attributes ([#14864](https://github.com/Lightning-AI/lightning/pull/14864))
 
+
 - Removed the deprecated `Trainer.verbose_evaluate`
+
+
+- Remove the deprecated `Trainer.should_rank_save_checkpoint` ([#14885](https://github.com/Lightning-AI/lightning/pull/14885))
 
 
 ### Fixed